# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: 群想 AI 群聊 API
  version: 1.0.0
  description: |
    以 backend/src 真实路由为准的接口契约。覆盖认证、短信、群组、消息、文件、
    TTS、社交互动、记忆、Agent、互动分析、监控、画像、人设与用户 API 配置等域。
    部分复杂查询端点（搜索、互动分析、监控指标等）的响应体为宽松对象，
    字段以实现为准，后续逐步收紧。
servers:
  - url: http://localhost:3002
    description: 本地后端
  - url: https://{YOUR_BACKEND_HOST}
    description: 生产后端（占位符，替换为真实 Render/自有域名）
    variables:
      YOUR_BACKEND_HOST:
        default: your-backend-host.example.com
...
security:
  - cookieAuth: []
  - csrfHeader: []
paths:
...
  /api/social/evaluate-like:
    post:
      tags: [Social]
      operationId: evaluateLike
      summary: 智能点赞评估（单条）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmartLikeInput'
      responses:
        '200':
          description: 评估结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/social/auto-like:
    post:
      tags: [Social]
      operationId: autoLike
      summary: 对指定消息执行自动点赞
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoLikeInput'
      responses:
        '200':
          description: 点赞执行结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/social/batch-evaluate:
    post:
      tags: [Social]
      operationId: batchEvaluate
      summary: 批量评估消息并给出推荐
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 批量评估结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/social/stats:
    get:
      tags: [Social]
      operationId: getSocialStats
      summary: 社交互动统计
      parameters:
        - in: query
          name: timeRange
          schema: { type: string }
      responses:
        '200':
          description: 统计数据
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/social/top-messages:
    get:
      tags: [Social]
      operationId: getTopMessages
      summary: 热门消息榜
      parameters:
        - in: query
          name: metric
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: 热门消息
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/social/active-participants:
    get:
      tags: [Social]
      operationId: getActiveParticipants
      summary: 活跃参与者
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: 活跃参与者列表
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/social/smart-like-config:
    get:
      tags: [Social]
      operationId: getSmartLikeConfig
      summary: 获取智能点赞配置
      responses:
        '200':
          description: 当前配置
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      tags: [Social]
      operationId: updateSmartLikeConfig
      summary: 更新智能点赞配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 配置更新结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/social/reset:
    post:
      tags: [Social]
      operationId: resetSocialData
      summary: 重置社交数据（管理员）
      responses:
        '200':
          description: 重置结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/social/comments/analyze:
    post:
      tags: [Social]
      operationId: analyzeComments
      summary: 评论语义分析
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 分析结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/social/comments/suggestions:
    get:
      tags: [Social]
      operationId: getSuggestedReplies
      summary: 获取评论回复建议
      parameters:
        - in: query
          name: messageId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 建议列表
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/social/comments/tree:
    get:
      tags: [Social]
      operationId: getCommentTree
      summary: 获取消息评论树（含深度校验）
      parameters:
        - in: query
          name: messageId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 评论树
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'

  # ------------------------------------------------------------------ Memory
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: 登录端点下发的 httpOnly 会话 Cookie
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
      description: |
        所有非安全方法（POST/PUT/PATCH/DELETE）必须携带，
        令牌通过 GET /api/csrf-token 获取，且须与 Cookie 中的
        CSRF 值一致（double-submit 模式）
...
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Forbidden:
      description: 无权访问（CSRF 校验失败/权限不足）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SmartLikeInput:
      type: object
      required: [message, contextMessages, senderInfo]
      properties:
        message:
          type: object
          additionalProperties: true
        contextMessages:
          type: array
          items:
            type: object
            additionalProperties: true
        senderInfo:
          type: object
          additionalProperties: true
    AutoLikeInput:
      type: object
      required: [messageId, groupId]
      additionalProperties: false
      properties:
        messageId: { type: string }
        groupId: { type: string }
...